day17/part1: fix leave trace (still not working D:)
Hamster traces are dependent on the direction they face, as well as the
amount of straight steps they already took, because it affects how they
can move in the future.

Example: Hamster A leaves a trace, but has already walked straight
three times, so in the future, it can't walk straight anymore. Hamster B
encounters the trace of Hamster A, but can still walk straight (it can
drink more beer than Hamster A). Even if Hamster B's trace so far is
worse than Hamster A's, it can still beat A, as it is still able to walk
straight. (Same is true for direction)