day17/part2: final fix
In part1, the hamster's entrance direction did not matter, as it would
immediately orient() and thus a hamster for each direction would exist.

With the introduction of the min straight movement threshold, this is no
longer possible. So at the beginning, it requires to setup a hamster for
each possible direction.